lvm2: Implement 'scalable mode' for large number of LVs in the system
For consistency reasons we would like to run updates sync, however
with large number of LVs in the system this can take considerable
amount of time, blocking the serialized daemon uevent processing.

As a workaround this change introduces an arbitrary threshold
that will switch updates back to async as was historically observed.