udiskslinuxprovider: Reorder uevent processing for modules
This is a major change and will need to be followed up to make
things right. This essientially reorders the moment the uevent
is propagated to modules and may affect availability of information
on exposed D-Bus objects.

This is all about dependencies, i.e. in case a module gathers
information from existing block objects, it needs to know whether
that block object has already finished processing the particular
uevent or not. This commit defines that order and it appears to be
sensible to guarantee the "core" daemon objects and its interfaces
are updated by the time udisks_module_new_object() is called. That
happens as a last step of uevent processing chain which includes
block object specific module callbacks for exposing extra D-Bus
interfaces (run prior to udisks_module_new_object()).

However it may not suit all cases and would be perhaps wise to provide
pre- and post- udisks_module_new_object() callbacks for finer
granularity and precise definition of the moment of uevent processing.

For the moment let's run udisks_module_new_object() as the last
in the chain and make modules aware of that.