udiskslinuxdriveata:Fix reading IO stats on ata_refresh_smart_sync
When checking IO stats after get_pm_state the noio bool will never become true
because get_pm_state() reads from the disk. If we perform the check earlier and
there was no IO on the disk, noio will become true and smart update refresh will be canceled.

This fixes the problem with StandbyTimeout not working if the timeout is bigger then 10 minutes.