udisksctl: fix autocomplete for power-off command
In order to autocomplete the "power-off" command we need to:

  1. Search for all dbus objects associated to the "org.freedesktop.UDisks2"
     service and...

  2. ...only keep (and show as autocompletion possibilities) the ones
     than meet "certain criteria".

That "certain criteria" is what this patch changes.

BEFORE this patch the filtering criteria was:
  - Objects that implement the "org.freedesktop.UDisks2.Drive.Ata" interface.

AFTER this patch the new criteria is:
  - Objects that implement the "org.freedesktop.UDisks2.Drive" interface and...
  - ...its "CanPowerOff()" method returns true.

Was the old criteria a bug? I don't know... but my USB dongles now
appear in the autocomplete list and didn't before this patch.