udiskslinuxpartitiontable: Rework block device locking
Nowadays with synthetic uevent tagging in use partitioning
can become more predictable. The right approach should be
to acquire an advisory lock, make changes, unlock and trigger
re-read.

Changing the shared lock to exclusive seems to make a big
difference and the underlying libblockdev-3 partitioning copes
with it just fine.