tests: Add basic NVMe tests
There's a known limitation in kernels prior to 5.18 where uevents are
not generated for controller state changes. This presents a race condition
with fabrics connections not always ready right after udev has added their
device nodes. An explicit udevadm trigger is added for such case but
timing is critical here and udisksd may not register all events within
the 1-second sleep added in the tests.