tests: Reflect fs signature vs. fs driver changes
The optional 'fstype' argument of the org.freedesktop.UDisks2.Filesystem.Mount()
method was always intended to act as an override. Now with the extension
of configurable mount options it defeats its purpose and is mostly
unnecessary in the regular mount tests. Remains to be used for
configurable mount options tests as we want to test multiple scenarios
and overrides.