udiskslinuxnvmecontroller: Check capabilities for self-test and sanitize log retrieval
In an ideal world the calls to the Get Log Page admin command for a log page
that is unsupported by the controller would result in a respective error
returned and the caller would accept this fate. Unfortunately in the real
world the kernel driver also logs pretty severe error message:

  nvme0: Admin Cmd(0x2), I/O Error (sct 0x0 / sc 0x2) DNR

While this can be safely ignored, such cryptic message printed in red
wouldn't make users happy. The NVMe Identify Controller structure exposes
a couple of flags that indicates feature support for some, though quite
limited at the end.

The NVMe 2.0 specification indicates mandatory support for the SMART
/ Health Information Log Page at the controller scope so the changes
in the commit are limited only to self-test and sanitize log pages and
operations.