udiskslinuxnvmenamespace: Avoid retrieving namespace info when format progress reporting is unsupported
The Format Namespace job periodically attempts to retrieve namespace info
despite the only information needed is the current format progress value.
Avoid this call unless progress reporting is indicated as supported.

On some drives when format is in progress the Identify Namespace admin
command may block until the operation is finished only to be cancelled
due to controller reset.