iscsi: Document the need of iscsid service restart after changing InitiatorName
The SetInitiatorName() method currently only writes the new string
to the /etc/iscsi/initiatorname.iscsi file yet the iscsid service needs
to be restarted to reflect the change. As this is highly distribution
specific and comes with issues (cgroups, security contexts, etc.),
it is not performed by udisksd.

Therefore, add a note to the documentation.