daemon: Remove UDisksObject interfaces more atomically
When an interface is being unexported and removed from the object,
other threads may be accessing its pointer that has not yet
been set to NULL.

This change will set it to NULL early in a slightly more atomic
way before continuing with destroying it. This may possibly increase
chance of segfaults (accessing NULL interface pointer) until
this is properly fixed.

See #1048