Extend Filesystem.Mount() to allow mounting on behalf of another user
Filesystems are mounted for the user who makes the D-Bus API call.
This has an effect on a few things, notably who has permissions
to unmount it later and also the path of the mount point, usually in
the form of /run/media/USER/LABEL

This patch adds a new parameter to the Filesystem.Mount() method that
allows mounting a filesystem on behalf of a different user.

Fixes #1065