tests: Skip protective parttable tests if none was created
Although a protective partition table is properly created on the disk,
its detection by kernel is rather random. Haven't found a way to force
probing and its presence, issuing partprobe on the device usually hides
it at all. So just make the test to work with what's available and skip
if nothing has been detected.