Limit getting filesystem size only to Ext and XFS
We switched to using bd_fs_get_size in 2.10 which added support
for getting size on NTFS, exFAT, VFAT and btrfs, which is already
causing issues, we should limit the filesystem where we try to
get the size manually using the filesystems tools to Ext and XFS
which we already did in 2.9. We can still get the data from udev
where possible.