lvm: Also look at symlinks for segment PVs
When a PV is on a LUKS device, lvm reports that device as
"/dev/mapper/luks-1234...", which is a symlink to the real device.
UDisks2 would miss this and report the segment as having lost its
backing PV.

In addition to looking harder for devices, UDisks2 now also
distinguises between "PV is gone" and "PV is there but I can't find a
D-Bus object for it".