udiskslinuxdrive: Calculate drive size from all attached NVMe namespaces
In case capacity reporting is not supported by the NVMe controller,
calculate the drive size from the currently attached namespaces
as a workaround.

This presents a chicken-egg problem when not all block objects are
present at the time the drive size calculation is done. So ping
the drive object back once a namespace interface is published,
and on all subsequent uevents (e.g. as a result of LBA format
change). Since the pingback is queued in the main loop, the drive
Size property gets updates slightly later.