integration-test: Fix change label test with exfatprogs 1.2.5
Like VFAT, EXFAT does not allow some characters in the label.
Since version 1.2.5 exfatprogs refuses to set a label with those
invalid characters.
Because this changed in version 1.2.5 we cannot assert that setting
such label would throw without knowing the exfatprogs version.

While we're touching this part of the test, also document why we're
using a shorter label compared to the other filesystems.