udiskslinuxmdraid: Avoid putting inhibit lock for externally started operations
UDisks publishes a new job object whenever it sees an array in sync,
repair or generally any other operation running. However this involves
acquiring and holding a system inhibit lock for the whole duration
of the operation, which for large disks may take hours or days.
Besides, mdraid resync is a restartable operation, defeating the purpose
for an inhibit lock.

This is also causing problems for installation environments where
UDisks is blocking post-installation reboot until the operation is done.