udisksstate: Always allocate store buffer for empty variants
The recent glib change added a g_variant_store() safeguard checking
for non-NULL buffer:
https://gitlab.gnome.org/GNOME/glib/-/commit/39c05b13123b12622ee2c93c170dbf20b573f6ac

For empty variants, their size is zero and g_variant_store() doesn't
actually serialize anything in the buffer. This used to work for ages
and a NULL buffer pointer was simply ignored and unused.

Work this around by always allocating a buffer even though zero
bytes are then stored in the file.