Fix using label as DM name when unlocking encrypted devices
'/' which is allowed in labels is not allowed in DM names so this
replaces '/' with underscore and also makes sure the name is
shorter than 128 characters.

See also #1404 where we disabled using labels for the unlocked
device name for BitLocker because BitLocker labels have slashes
by default (including our test image).