udiskslinuxdriveobject: Hold interface references during housekeeping
As long as housekeeping is run in a thread, it needs to hold interface
GObject references for the whole time it runs. While uevent processing
is done in a single thread and DriveObject interface pointers are
exchanged atomically, they're still referenced after being freed.

Surprisingly this has mostly worked in practice and only led to
a very small amount of crashes that were not obvious on a first sight.