udiskslinuxprovider: Filter invalid NVMe Subsystem NQN values out
In certain situations, e.g. when the kernel host nvme driver failed
to probe the device or when the controller was not in a 'live' state,
an invalid Subsystem NQN values may appear in sysfs. In such a case
subsystem tracking should be skipped with an expectation of another
uevent once the controller recovers or becomes live.

The commonly seen values are '(efault)' and '(null)'. See also
https://www.kernel.org/doc/Documentation/core-api/printk-formats.rst